+++ test/test_helper.exs
@@ -1,3 +1,7 @@
 Mox.defmock(X402.PaymentIdentifierCacheMock, for: X402.Extensions.PaymentIdentifier.Cache)
 
-ExUnit.start(exclude: [:smoke])
+Mox.defmock(X402.RedisCommandMock,
+  for: X402.Extensions.PaymentIdentifier.RedisCache.Command
+)
+
+ExUnit.start(exclude: [:smoke, :redis])